requirements.txt
### dependency
streamlit
openai
langchain
langchain_community
arxiv
pymupdf
langchain-openai
python-dotenv